tfm: validate and fix next larger programs
This validation and fixing code is invoked both in pltotf and tftopl, and the
E2E tests verify both cases.

+++ crates/tfm/src/ligkern/lang.rs
@@ -426,7 +426,7 @@ impl ValidationWarning {
     }
 
     /// Returns the section in Knuth's TFtoPL (version 2014) in which this warning occurs.
-    pub fn tftopl_section(&self) -> usize {
+    pub fn tftopl_section(&self) -> u8 {
         use ValidationWarning::*;
         match self {
             SkipTooLarge(_) => 70,
